Update to scala 2.11
- Mostly just replaced deprecated functions and make functions protected
  where necessary.

- Add a fix to allow @transient to work on constructor parameters.

- Fixes for many namespace changes. It appears that the scala-xml module
  fixed a handful of bugs related to namespaces. This broke the
  DaffodilConstructingLoader, because we weren't correctly modifying
  scopes to include dafint (we added the scope bottom up rather than top
  down, which meant that child elements weren't properly sharing the
  dafint namespace binding with their parents, so every element had what
  appeared to be a unique dafint namespace binding. This was why when
  looking at a schema dump you would see the xmlns:dafint namespace
  declared on every element. This changes the way we construct those to
  build the namespace properly. This fixed a handful of namespace
  issues that cropped up with 2.11.

- Pull in changes from scala-xml in the PrettyPrinter that change the
  default top scope from null to TopScope. Without these fixes, the
  namespace xmlns="" was incorrectly added in some cases.

- Add the -Xxml:coalescing compiler flag, which prevents scala-xml from
  coalescing nodes. This allows us to keep the PCData information around
  and properly coalesce for our needs.

- Fixes removal of all whitespace text nodes. We don't actually want
  that. Now that PCData and Text nodes remain separate, it's possible
  that we have whitespace text nodes inbetween PCData nodes. This is
  perfectly acceptable, and we do not want to remove those whitespace
  nodes. What we really want to do is remove white space nodes that are
  mixed with Elem nodes. This changes XMLUtils to perform that action.

DFDL-1211

+++ daffodil-runtime1/src/main/scala/edu/illinois/ncsa/daffodil/dpath/FNFunctions.scala
@@ -332,7 +332,7 @@ case class FNRoundHalfToEven(recipeNum: CompiledDPath, recipePrecision: Compiled
     val bd = unrounded match {
       case s: String => BigDecimal(s) // TODO: Remove eventually. Holdover from JDOM where everything is a string.
       case l: Long => BigDecimal.valueOf(l)
-      case f: Float => BigDecimal.valueOf(f)
+      case f: Float => BigDecimal.valueOf(f.toDouble)
       case d: Double => BigDecimal.valueOf(d)
       case bd: BigDecimal => bd
       case _ => Assert.invariantFailed("not a number")
